fix: Fix issue #25 the sign up render bug
The sign up page was being briefly rendered whenever anypage was loaded.
The issue was the sign up page was trying to write directly to the DOM
root. It has been fixed.
