--- frontend/src/pages/signupApp.tsx
@@ -1,86 +1,0 @@
-// Import necessary libraries
-import React, { useState } from 'react';
-import axios from 'axios';
-
-function App() {
-  const [formData, setFormData] = useState({
-    username: '',
-    email: '',
-    password: '',
-    confirmPassword: ''
-  });
-
-  const handleChange = (e) => {
-    setFormData({ ...formData, [e.target.name]: e.target.value });
-  };
-
-  const handleSubmit = async (e) => {
-    e.preventDefault();
-
-    if (formData.password !== formData.confirmPassword) {
-      alert('Passwords do not match.');
-      return;
-    }
-
-    try {
-      const response = await axios.post(
-        'http://localhost:3001/register',
-        formData
-      );
-      alert(response.data.message);
-    } catch (error) {
-      alert('Registration failed. User may already exist.');
-    }
-  };
-
-  return (
-    <div className='container'>
-      <h1>Bookwormer Registration</h1>
-      <form onSubmit={handleSubmit}>
-        <label htmlFor='username'>Username:</label>
-        <input
-          type='text'
-          id='username'
-          name='username'
-          value={formData.username}
-          onChange={handleChange}
-          required
-        />
-        <br />
-        <label htmlFor='email'>Email:</label>
-        <input
-          type='email'
-          id='email'
-          name='email'
-          value={formData.email}
-          onChange={handleChange}
-          required
-        />
-        <br />
-        <label htmlFor='password'>Password:</label>
-        <input
-          type='password'
-          id='password'
-          name='password'
-          value={formData.password}
-          onChange={handleChange}
-          required
-        />
-        <br />
-        <label htmlFor='confirmPassword'>Confirm Password:</label>
-        <input
-          type='password'
-          id='confirmPassword'
-          name='confirmPassword'
-          value={formData.confirmPassword}
-          onChange={handleChange}
-          required
-        />
-        <br />
-        <button type='submit'>Sign Up</button>
-      </form>
-    </div>
-  );
-}
-
-export default App;
